pyproject.toml
[project]
name = "url-shortner"
version = "0.1.0"
description = "URL shortener with admin and site"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.2",
    "django-choices>=2.0.0",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "psycopg2-binary>=2.9.10",
]


[tool.ruff]
line-length = 110
exclude = ["migrations", "templates", "docs"]

[tool.ruff.lint]
extend-select = ["I"]
select = ["E", "F", "W"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[dependency-groups]
dev = [
    "django-silk>=5.4.0",
]
